+++ packages/sonataflow-management-console-webapp/src/components/console/ManagementConsoleRoutes/ManagementConsoleRoutes.tsx
@@ -18,14 +18,13 @@
  */
 import * as React from "react";
 import { Redirect, Route, Switch } from "react-router-dom";
-import EmptyPage from "../../pages/EmptyPage";
-import { RuntimeToolsWorkflowDefinitions } from "../../../runtimeTools/pages/RuntimeToolsWorkflowDefinitions";
 import { routes } from "../../../navigation/Routes";
-import { RuntimeToolsWorkflowForm } from "../../../runtimeTools/pages/RuntimeToolsWorkflowForm";
-import { RuntimeToolsWorkflowInstances } from "../../../runtimeTools/pages/RuntimeToolsWorkflowInstances";
-import { RuntimeToolsWorkflowDetails } from "../../../runtimeTools/pages/RuntimeToolsWorkflowDetails";
-import { RuntimeToolsTriggerCloudEvent } from "../../../runtimeTools/pages/RuntimeToolsTriggerCloudEvent";
+import { TriggerCloudEventPage } from "../../pages/TriggerCloudEventPage/TriggerCloudEventPage";
 import MonitoringPage from "../../pages/MonitoringPage/MonitoringPage";
+import { WorkflowDefinitionsPage } from "../../pages/WorkflowDefinitionsPage/WorkflowDefinitionsPage";
+import { WorkflowInstancesPage } from "../../pages/WorkflowInstancesPage/WorkflowInstancesPage";
+import { WorkflowDetailsPage } from "../../pages/WorkflowDetailsPage/WorkflowDetailsPage";
+import { WorkflowFormPage } from "../../pages/WorkflowFormPage/WorkflowFormPage";
 
 const ManagementConsoleRoutes: React.FC = () => {
   return (
@@ -35,14 +34,14 @@ const ManagementConsoleRoutes: React.FC = () => {
         path={routes.home.path({})}
         render={() => <Redirect to={routes.runtimeToolsWorkflowInstances.path({})} />}
       />
-      <Route exact path={routes.runtimeToolsWorkflowInstances.path({})} component={RuntimeToolsWorkflowInstances} />
-      <Route exact path={routes.runtimeToolsWorkflowDefinitions.path({})} component={RuntimeToolsWorkflowDefinitions} />
+      <Route exact path={routes.runtimeToolsWorkflowInstances.path({})} component={WorkflowInstancesPage} />
+      <Route exact path={routes.runtimeToolsWorkflowDefinitions.path({})} component={WorkflowDefinitionsPage} />
       <Route path={routes.runtimeToolsWorkflowDetails.path({ workflowId: ":workflowId" })}>
-        {({ match }) => <RuntimeToolsWorkflowDetails workflowId={match!.params.workflowId!} />}
+        {({ match }) => <WorkflowDetailsPage workflowId={match!.params.workflowId!} />}
       </Route>
-      <Route exact path={routes.runtimeToolsTriggerCloudEvent.path({})} component={RuntimeToolsTriggerCloudEvent} />
+      <Route exact path={routes.runtimeToolsTriggerCloudEvent.path({})} component={TriggerCloudEventPage} />
       <Route path={routes.runtimeToolsWorkflowForm.path({ workflowName: ":workflowName" })}>
-        <RuntimeToolsWorkflowForm />
+        <WorkflowFormPage />
       </Route>
       <Route exact path={routes.monitoring.path({})} component={MonitoringPage} />
     </Switch>
